# docs/openapi.yaml
openapi: 3.0.3
info:
  title: TotHub Daycare Management API
  description: |
    Enterprise-grade childcare management platform API for attendance tracking, 
    staff scheduling, compliance monitoring, and parent communication.
    
    ## Key Features
    - Multi-tenant architecture with role-based access control
    - Real-time attendance and ratio monitoring
    - Multi-state compliance rules engine
    - Staff scheduling and payroll management
    - Parent portal and communication tools
    - Biometric authentication support
    - QuickBooks integration
    
    ## Authentication
    All API calls require a valid JWT token in the Authorization header:
    `Authorization: Bearer <token>`
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests/hour
    - Authentication endpoints: 100 requests/hour
    - File uploads: 100 requests/hour
  
  version: 1.0.0
  contact:
    name: TotHub Development Team
    email: dev@tothub.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.tothub.com/v1
    description: Production server
  - url: https://staging-api.tothub.com/v1
    description: Staging server
  - url: http://localhost:5000/v1
    description: Local development
security:
  - bearerAuth: []
paths:
...
  /compliance/check:
    post:
      summary: Check room compliance
      description: Comprehensive room compliance check including ratios, qualifications, and facility requirements
      tags: [Compliance]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - state
                - room
                - children
                - staff
              properties:
                state:
                  type: string
                  pattern: '^[A-Z]{2}$'
                  example: "WV"
                room:
                  $ref: '#/components/schemas/Room'
                children:
                  type: array
                  items:
                    $ref: '#/components/schemas/Child'
                staff:
                  type: array
                  items:
                    $ref: '#/components/schemas/Staff'
      responses:
        '200':
          description: Room compliance check result
          content:
            application/json:
              schema:
                type: object
                properties:
                  roomId:
                    type: string
                    format: uuid
                  state:
                    type: string
                  compliant:
                    type: boolean
                  violations:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComplianceViolation'
                  warnings:
                    type: array
                    items:
                      $ref: '#/components/schemas/ComplianceWarning'
                  complianceScore:
                    type: number
                    minimum: 0
                    maximum: 100
                  recommendations:
                    type: array
                    items:
                      type: string
                  timestamp:
                    type: string
                    format: date-time
        '400':
          description: Invalid request data
        '500':
          description: Internal server error
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        parentEmail:
          type: string
          format: email
        parentPhone:
          type: string
        emergencyContact:
          type: string
        medicalNotes:
          type: string
        enrollmentDate:
          type: string
          format: date
        isActive:
          type: boolean
        roomId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
...
    ComplianceViolation:
      type: object
      properties:
        type:
          type: string
          enum: [ratio, qualification, facility, safety]
        severity:
          type: string
          enum: [low, medium, high, critical]
        message:
          type: string
        roomId:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        details:
          type: object
    ComplianceWarning:
      type: object
      properties:
        type:
          type: string
          enum: [ratio_warning, qualification_warning, facility_warning]
        message:
          type: string
        roomId:
          type: string
          format: uuid
        staffId:
          type: string
          format: uuid
        details:
          type: object
    Room:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        tenantId:
          type: string
          format: uuid
        capacity:
          type: integer
          minimum: 1
        ageGroup:
          type: string
          enum: [Infant, Toddler, Preschool, School Age]
        isActive:
          type: boolean
    Staff:
      type: object
      properties:
        id:
          type: string
          format: uuid
        firstName:
          type: string
        lastName:
          type: string
        position:
          type: string
        qualifications:
          type: array
          items:
            type: string
        certifications:
          type: array
          items:
            type: string
        experienceYears:
          type: integer
          minimum: 0
        roomId:
          type: string
          format: uuid
        tenantId:
          type: string
          format: uuid
        isActive:
          type: boolean
          minimum: 0
        currentChildren:
          type: integer
          minimum: 0
        maxRatio:
          type: string
          example: "1:10"
        isCompliant:
          type: boolean
        lastUpdated:
          type: string
          format: date-time
...